Fix /compact command.
